# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PUCPR Projeto Final - Webservices e MBaaS ",
        "description": "API criada como projeto final da matéria de Webservices e MBaaS - por Lucas Cândido dos Santos",
        "contact": {
            "name": "Lucas Cândido dos Santos"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/destaques": {
            "get": {
                "tags": [
                    "produtos"
                ],
                "summary": "Retorna uma lista de produtos em destaque",
                "description": "Retorna uma lista de produtos em destaque",
                "operationId": "286debe8b038e5b45d8fb173f67249eb",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                }
            }
        },
        "/pesquisa/{where}": {
            "get": {
                "tags": [
                    "produtos"
                ],
                "summary": "Retorna uma lista de produtos baseados no termo pesquisado",
                "description": "Retorna uma lista de produtos baseados no termo pesquisado",
                "operationId": "3cf8252be6c9e66e6831d2782ae7bf73",
                "parameters": [
                    {
                        "name": "where",
                        "in": "path",
                        "description": "Opções de busca -> busca=shampoo; busca=todos; marca=123; categoria=456",
                        "required": true,
                        "example": "busca=todos"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                }
            }
        },
        "/produtos/{id}": {
            "get": {
                "tags": [
                    "produtos"
                ],
                "summary": "Retorna um produto baseado no ID informado",
                "description": "Retorna um produto baseado no ID informado",
                "operationId": "04ff89b620ba602adef63569d15d944d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "example": "123"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Missing Data"
                    }
                }
            },
            "put": {
                "tags": [
                    "produtos"
                ],
                "summary": "Atualiza um produto",
                "description": "Atualiza um produto",
                "operationId": "c45ef52a7b159513ec0c8878561a6c38",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "example": "123"
                    },
                    {
                        "name": "produto",
                        "in": "query",
                        "description": "Código de referencia do produto",
                        "required": false
                    },
                    {
                        "name": "nome",
                        "in": "query",
                        "description": "Nome do produto",
                        "required": false
                    },
                    {
                        "name": "quantidade",
                        "in": "query",
                        "description": "Quantidade em estoque do produto",
                        "required": false
                    },
                    {
                        "name": "preco",
                        "in": "query",
                        "description": "Preço de venda do produto",
                        "required": false
                    },
                    {
                        "name": "tamanho",
                        "in": "query",
                        "description": "Tamanho do produto",
                        "required": false
                    },
                    {
                        "name": "cor",
                        "in": "query",
                        "description": "Cor do produto",
                        "required": false
                    },
                    {
                        "name": "inativo",
                        "in": "query",
                        "description": "Flag que define se o produto está ativo ou não",
                        "required": false
                    },
                    {
                        "name": "imagem",
                        "in": "query",
                        "description": "URL da imagem do produto",
                        "required": false
                    },
                    {
                        "name": "destaque",
                        "in": "query",
                        "description": "Flag que define se um produto é destaque ou não",
                        "required": false
                    },
                    {
                        "name": "preco_promocional",
                        "in": "query",
                        "description": "Preço promocional do produto",
                        "required": false
                    },
                    {
                        "name": "descricao",
                        "in": "query",
                        "description": "Descrição do produto",
                        "required": false
                    },
                    {
                        "name": "marca_id",
                        "in": "query",
                        "description": "ID da marca",
                        "required": false
                    },
                    {
                        "name": "categoria_id",
                        "in": "query",
                        "description": "ID da categoria",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Não autorizado. Apenas administradores podem cadastrar produtos"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}